Final project refs (#2)
* finalize project refs

* update tsconfigs

* get back deps that are crashing build

* get back all the deps and fix all possible errors in plugin-core

* remove unused thing

+++ packages/api-server/src/utils.ts
@@ -84,7 +84,9 @@ export class ServerUtils {
     const handleExit = () => {
       console.log("handle exit");
       try {
-        process.kill(subprocess.pid);
+        if (subprocess.pid) {
+          process.kill(subprocess.pid);
+        }
       } catch (err: any) {
         // this means process was already killed
         if (err.code !== "ESRCH") {
